punto2_api/requirements.txt
fastapi
uvicorn
sqlalchemy>=2
psycopg2-binary
pydantic>=2
python-dotenv
PyJWT
pandas
openai